@@ -7,8 +7,30 @@ import org.json.JSONObject
 import java.io.Serializable
 import java.util.ArrayList
 
-class Pokemon : Serializable {
-    var id: Int = 0
+
+class Pokemon(
+        var name: String?,
+        var sprites: Sprites?,
+        var id: Int?,
+        var abilities: List<Abilities>?,
+        var types: List<Types>?,
+        var index: Int): Serializable
+
+object PokeDex {
+    val pokeDex = mutableListOf<Pokemon>()
+}
+
+class Ability(val name: String)
+
+class Abilities(val ability: Ability)
+
+class Type(val name: String)
+
+class Types(val type: Type)
+
+class Sprites(val front_shiny: String)
+
+    /*var id: Int = 0
         private set
     var moves: ArrayList<String>? = null
         private set
@@ -69,5 +91,5 @@ class Pokemon : Serializable {
             e.printStackTrace()
         }
 
-    }
-}
+    }*/
+
